Remove the SQLite namespace
