+++ README.md
@@ -1,4 +1,4 @@
-# SQLite.Database
+# Database
 
 [![Swift](https://img.shields.io/badge/swift-5.2-green.svg?longCache=true&style=flat)](https://developer.apple.com/swift/)
 [![License](https://img.shields.io/badge/license-MIT-green.svg?longCache=true&style=flat)](/LICENSE)
@@ -6,7 +6,7 @@
 
 ## Introduction
 
-[SQLite.Database]() is a simple Swift wrapper around [SQLite](http://www.sqlite.org/). It is intended to act as an introduction to SQLite for Swift developers. You can read more about that [here](https://shareup.app/blog/building-a-lightweight-sqlite-wrapper-in-swift/). SQLite.Database allows developers to use pure SQL to access or modify their database without forcing them to deal with all the tiresome minutia involved in configuring SQLite databases and queries or converting from Swift types to the C types that SQLite expects.
+[Database]() is a simple Swift wrapper around [SQLite](http://www.sqlite.org/). It is intended to act as an introduction to SQLite for Swift developers. You can read more about that [here](https://shareup.app/blog/building-a-lightweight-sqlite-wrapper-in-swift/). Database allows developers to use pure SQL to access or modify their database without forcing them to deal with all the tiresome minutia involved in configuring SQLite databases and queries or converting from Swift types to the C types that SQLite expects.
 
 ## Installation
 
@@ -24,7 +24,7 @@ let package = Package(
 
 ## License
 
-The license for SQLite.Database is the standard MIT licence. You can find it in the `LICENSE` file.
+The license for Database is the standard MIT licence. You can find it in the `LICENSE` file.
 
 ## Alternatives
