+++ Sources/SQLite/Encoder.swift
@@ -1,44 +1,42 @@
 import Foundation
 
-extension SQLite {
-    public final class Encoder {
-        public enum Error: Swift.Error {
-            case invalidType(Any)
-            case invalidValue(Any)
-            case invalidJSON(Data)
-            case transactionFailed
-        }
+public final class Encoder {
+    public enum Error: Swift.Error {
+        case invalidType(Any)
+        case invalidValue(Any)
+        case invalidJSON(Data)
+        case transactionFailed
+    }
 
-        private let _database: SQLite.Database
+    private let _database: Database
 
-        public init(_ database: SQLite.Database) {
-            _database = database
-        }
+    public init(_ database: Database) {
+        _database = database
+    }
 
-        public func encode<T: Encodable>(_ value: T, using sql: SQL) throws {
-            let encoder = _SQLiteEncoder()
+    public func encode<T: Encodable>(_ value: T, using sql: SQL) throws {
+        let encoder = _SQLiteEncoder()
 
-            if let array = value as? Array<Encodable> {
-                let success = try _database.inTransaction {
-                    try array.forEach { (element: Encodable) in
-                        try element.encode(to: encoder)
-                        try _database.write(sql, arguments: encoder.encodedArguments)
-                    }
+        if let array = value as? Array<Encodable> {
+            let success = try _database.inTransaction {
+                try array.forEach { (element: Encodable) in
+                    try element.encode(to: encoder)
+                    try _database.write(sql, arguments: encoder.encodedArguments)
                 }
-                if success == false {
-                    throw SQLite.Encoder.Error.transactionFailed
-                }
-            } else if let dictionary = value as? Dictionary<AnyHashable, Encodable> {
-                throw SQLite.Encoder.Error.invalidType(dictionary)
-            } else {
-                try value.encode(to: encoder)
-                try _database.write(sql, arguments: encoder.encodedArguments)
             }
+            if success == false {
+                throw Encoder.Error.transactionFailed
+            }
+        } else if let dictionary = value as? Dictionary<AnyHashable, Encodable> {
+            throw Encoder.Error.invalidType(dictionary)
+        } else {
+            try value.encode(to: encoder)
+            try _database.write(sql, arguments: encoder.encodedArguments)
         }
     }
 }
 
-private class _SQLiteEncoder: Encoder {
+private class _SQLiteEncoder: Swift.Encoder {
     var codingPath: Array<CodingKey> = []
     var userInfo: [CodingUserInfoKey : Any] = [:]
 
@@ -115,7 +113,7 @@ private struct _KeyedContainer<K: CodingKey>: KeyedEncodingContainerProtocol {
     }
 
     mutating func encode(_ value: UInt64, forKey key: K) throws {
-        guard value < Int64.max else { throw SQLite.Encoder.Error.invalidValue(value) }
+        guard value < Int64.max else { throw Encoder.Error.invalidValue(value) }
         _storage[key.stringValue] = .integer(Int64(value))
     }
 
@@ -136,7 +134,7 @@ private struct _KeyedContainer<K: CodingKey>: KeyedEncodingContainerProtocol {
     }
 
     mutating func encode(_ value: Date, forKey key: K) throws {
-        let string = SQLite.DateFormatter.string(from: value)
+        let string = PreciseDateFormatter.string(from: value)
         _storage[key.stringValue] = .text(string)
     }
 
@@ -160,7 +158,7 @@ private struct _KeyedContainer<K: CodingKey>: KeyedEncodingContainerProtocol {
         } else {
             let jsonData = try jsonEncoder.encode(value)
             guard let jsonText = String(data: jsonData, encoding: .utf8) else {
-                throw SQLite.Encoder.Error.invalidJSON(jsonData)
+                throw Encoder.Error.invalidJSON(jsonData)
             }
             _storage[key.stringValue] = .text(jsonText)
         }
@@ -174,11 +172,11 @@ private struct _KeyedContainer<K: CodingKey>: KeyedEncodingContainerProtocol {
         fatalError("_KeyedContainer does not support nested containers.")
     }
 
-    mutating func superEncoder() -> Encoder {
+    mutating func superEncoder() -> Swift.Encoder {
         fatalError("_KeyedContainer does not support nested containers.")
     }
 
-    mutating func superEncoder(forKey key: K) -> Encoder {
+    mutating func superEncoder(forKey key: K) -> Swift.Encoder {
         fatalError("_KeyedContainer does not support nested containers.")
     }
 }
@@ -192,7 +190,7 @@ private class _KeyedStorage {
         _elements.removeAll(keepingCapacity: true)
     }
 
-    subscript(key: String) -> SQLite.Value? {
+    subscript(key: String) -> SQLiteValue? {
         get {
             return _elements[key]
         }
@@ -206,7 +204,7 @@ private let jsonEncoder: JSONEncoder = {
     let encoder = JSONEncoder()
     encoder.dataEncodingStrategy = .base64
     encoder.dateEncodingStrategy = .custom({ (date, encoder) throws in
-        let dateAsString = SQLite.DateFormatter.string(from: date)
+        let dateAsString = PreciseDateFormatter.string(from: date)
         var container = encoder.singleValueContainer()
         try container.encode(dateAsString)
     })
