+++ Sources/SQLite/SQLite.swift
@@ -6,5 +6,3 @@ public typealias Statement = OpaquePointer
 
 let SQLITE_STATIC = unsafeBitCast(0, to: sqlite3_destructor_type.self)
 let SQLITE_TRANSIENT = unsafeBitCast(-1, to: sqlite3_destructor_type.self)
-
-public struct SQLite { }
